requirements.txt
openai
google-genai
python-dotenv
